+++ src/controls/select.ts
@@ -37,13 +37,14 @@ declare global {
       /** Case when menu-popup to show
        * @defaultValue onPressArrowKey | onClick | onFocus | onInput */
       showCase: ShowCases;
+      /** Set true to make input not editable but allow select items via popup-menu (ordinary dropdown mode)
+       * @tutorial
+       * * set number X to enable autoMode where input.readOnly = items.length < X */
+      readOnlyInput?: boolean | number;
     }
     interface Options<T = any, VM = ValidityMap> extends WUP.BaseCombo.Options<T, VM>, Defaults<T, VM> {
       /** Items showed in dropdown-menu. Provide promise/api-call to show pending status when control retrieves data! */
       items: MenuItems<T> | (() => MenuItems<T> | Promise<MenuItems<T>>) | Promise<MenuItems<T>>;
-      /** Set true to make input not editable but allow to user select items via popup-menu (ordinary dropdown mode)
-       *  @defaultValue false */
-      readOnlyInput?: boolean;
       /** Allow user to create new value if value not found in items
        * @defaultValue false */
       allowNewValue?: boolean;
@@ -284,6 +285,7 @@ export default class WUPSelectControl<
       this._onPendingInitValue?.call(this);
       delete this._onPendingInitValue;
       this.setInputValue(this.$value, SetValueReasons.clear);
+      this.setupInputReadonly(); // call it because opt readonlyInput can depends on items.length
     };
     if (d instanceof Promise) {
       return promiseWait(d, 300, (v) => this.changePending(v)).then((data) => {
@@ -313,9 +315,10 @@ export default class WUPSelectControl<
     super.gotChanges(propsChanged as any);
   }
 
-  override gotFormChanges(propsChanged: Array<keyof WUP.Form.Options> | null): void {
-    super.gotFormChanges(propsChanged);
-    this.$refInput.readOnly = this.$refInput.readOnly || (this.$isPending as boolean);
+  override setupInputReadonly(): void {
+    const r = this._opts.readOnlyInput;
+    this.$refInput.readOnly =
+      this.$isReadOnly || this.$isPending || r === true || (typeof r === "number" && r < this._cachedItems!.length);
   }
 
   override setupInitValue(propsChanged: Array<keyof WUP.Select.Options> | null): void {
@@ -792,4 +795,3 @@ customElements.define(tagName, WUPSelectControl);
 
 // NiceToHave: option to allow autoselect item without pressing Enter
 // WARN Chrome touchscreen simulation issue: touch on label>strong fires click on input - the issue only in simulation
-// todo setup readOnlyInput: 7 - where need to set readOnly only when items.length < 7